requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-APScheduler==1.12.3
Pillow==8.3.1
requests==2.26.0
python-dotenv==0.19.0
email-validator==1.1.3
WTForms==2.3.3
SQLAlchemy==1.4.23
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
gunicorn==20.1.0
pytest==6.2.5
pytest-flask==1.2.0